move sources to pkg/mindpak
